Time: 1 ms (71.01%), Space: 42 MB (31.67%) - LeetHub

+++ 1464-maximum-product-of-two-elements-in-an-array/1464-maximum-product-of-two-elements-in-an-array.java
@@ -1,7 +1,7 @@
 class Solution {
     public int maxProduct(int[] nums) {
         int m1=1,m2=1;
-        for (int i=0; i<nums.length; i++) {
+        for (short i=0; i<nums.length; i++) {
             if (nums[i]>m1) {
                 if (m2<m1) m2=m1; 
                 m1=nums[i];
